requirements.txt
# Core JAX/Flax dependencies
jax[tpu]>=0.4.23
flax>=0.8.0

# Data processing
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# ML utilities
scikit-learn>=1.0.0
optax>=0.1.7

# Logging and tracking
wandb>=0.15.0

# HuggingFace integration
huggingface-hub>=0.19.0
datasets>=2.14.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Progress bars
tqdm>=4.65.0